# Significantly faster offset cache for parser translation with multibyte characters