Before:

```
ruby 3.4.0dev (2024-02-07T11:34:48Z master 9ebaf7a8ef) [arm64-darwin23]
Warming up --------------------------------------
 Parser::CurrentRuby     1.000 i/100ms
Prism::Translation::Parser
                         1.000 i/100ms
Calculating -------------------------------------
 Parser::CurrentRuby      1.203 (± 0.0%) i/s -      7.000 in   5.821395s
Prism::Translation::Parser
                          0.245 (± 0.0%) i/s -      2.000 in   8.219212s

Comparison:
 Parser::CurrentRuby:        1.2 i/s
Prism::Translation::Parser:        0.2 i/s - 4.91x  slower
```

after:

```
ruby 3.4.0dev (2024-02-07T11:34:48Z master 9ebaf7a8ef) [arm64-darwin23]
Warming up --------------------------------------
 Parser::CurrentRuby     1.000 i/100ms
Prism::Translation::Parser
                         1.000 i/100ms
Calculating -------------------------------------
 Parser::CurrentRuby      1.167 (± 0.0%) i/s -      6.000 in   5.140792s
Prism::Translation::Parser
                          2.224 (± 0.0%) i/s -     12.000 in   5.401231s

Comparison:
Prism::Translation::Parser:        2.2 i/s
 Parser::CurrentRuby:        1.2 i/s - 1.90x  slower
```